Workaround due to bug in GHC. Refs #30.
There is a bug in 8 series of GHC that makes certain arrow expressions
not compile.

The bug has been detailed at:
https://gitlab.haskell.org/ghc/ghc/-/issues/15175
https://gitlab.haskell.org/ghc/ghc/-/issues/18950

The latter gives an alternative way of writing those expressions so that
they compile.

This commit re-writes the only expression in the game that triggers that
bug, and adds a comment to understand why it should not be changed (at
least not immediately).