Disable threaded RTS. Refs #33.
When using GHC 8.6.5, the execution is slow. It may also happen with
other versions of GHC, although it does not happen with GHC 7.10.3

When threaded is disabled, execution is fluid and fast. Since this game
has no threads or anything that needs to be executed in a parallel
(e.g., music playing), threaded is not needed.

This commit eliminates the -threaded flag from ghc-options in the cabal
file.