Update installation instructions. Refs #36.
The installation instructions mention a sandbox command that no longer
exists in newer versions of Cabal. They also contain some small
inconsistencies between the three snippets included in the README.

This commit updates the README with an installation procedure that works
with a recent (3.4) version of cabal-install.

It also modifies the comments and steps to be all as similar as possible
between them.

[ci skip]