Remove commented code. Refs #39.
The RenderObject module contains commented code.

Although some of this code could potentially and theoretically help
locally in the future, overall the accumulation will be more detrimental
to the maintenance effort of the project. Any comments worth mentioning
should be properly stated, either internally in the code or as part of
the haddock documentation.

This commit removes all commented code from the module.

[ci skip]